requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
Pillow==9.5.0
python-dotenv==1.0.0
celery==5.0
redis==3.5
